move POSTag class to data-definitions and replace old (incomplete) PosTag class
