+++ data-structures/src/main/java/edu/kit/kastel/mcse/ardoco/core/datastructures/definitions/IWord.java
@@ -8,7 +8,7 @@ public interface IWord {
 
     String getText();
 
-    PosTag getPosTag();
+    POSTag getPosTag();
 
     IWord getPreWord();
