--- data-structures/src/main/java/edu/kit/kastel/mcse/ardoco/core/datastructures/definitions/PosTag.java
@@ -1,8 +1,0 @@
-package edu.kit.kastel.mcse.ardoco.core.datastructures.definitions;
-
-public enum PosTag {
-
-    // TODO: To Complete
-
-    DT, NNP, NN, NNPS, NNS;
-}
